+++ pandas/core/indexes/datetimelike.py
@@ -605,7 +605,7 @@ class DatetimeTimedeltaMixin(DatetimeIndexOpsMixin, Int64Index):
     def factorize(self, sort=False, na_sentinel=-1):
         if self.freq is not None and sort is False:
             # we are unique, so can short-circuit, also can preserve freq
-            codes = np.arange(len(self))
+            codes = np.arange(len(self), dtype=np.intp)
             return codes, self[:]
             # TODO: In the sort=True case we could check for montonic_decreasing
             #  and operate on self[::-1]
